+++ pkg/local_object_storage/engine/gc_test.go
@@ -16,8 +16,10 @@ import (
 	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	objectSDK "github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
+	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
 	usertest "github.com/nspcc-dev/neofs-sdk-go/user/test"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/zap/zaptest"
 )
 
 func TestChildrenExpiration(t *testing.T) {
@@ -296,6 +298,90 @@ func TestGC(t *testing.T) {
 		// now GC must delete the object eventually
 		checkObjectsAsyncRemoval(t, e, cnr, obj.GetID(), tomb.GetID())
 	})
+}
+
+func TestSplitObjectExpirationWithoutLink(t *testing.T) {
+	dir := t.TempDir()
+	es := &asyncEpochState{e: 10}
+	l := zaptest.NewLogger(t)
+	e := New(WithLogger(l))
+	_, err := e.AddShard(
+		shard.WithLogger(l),
+		shard.WithBlobstor(
+			newStorage(filepath.Join(dir, "fstree")),
+		),
+		shard.WithMetaBaseOptions(
+			meta.WithLogger(l),
+			meta.WithPath(filepath.Join(dir, "metabase")),
+			meta.WithPermissions(0700),
+			meta.WithEpochState(es),
+		),
+		shard.WithExpiredObjectsCallback(e.processExpiredObjects),
+		shard.WithGCRemoverSleepInterval(100*time.Millisecond),
+	)
+	require.NoError(t, err)
+
+	require.NoError(t, e.Open())
+	require.NoError(t, e.Init())
+	t.Cleanup(func() { _ = e.Close() })
+
+	cnr := cidtest.ID()
+	parentID := oidtest.ID()
+
+	parent := generateObjectWithCID(cnr)
+	parent.SetID(parentID)
+	parent.SetPayload(nil)
+	addExpirationAttribute(parent, es.CurrentEpoch())
+	parentAddr := object.AddressOf(parent)
+
+	const childCount = 3
+	children := make([]*objectSDK.Object, childCount)
+	childIDs := make([]oid.ID, childCount)
+	childAddrs := make([]oid.Address, childCount)
+
+	var firstID oid.ID
+	for i := range children {
+		children[i] = generateObjectWithCID(cnr)
+		if i == 0 {
+			firstID = children[i].GetID()
+		}
+		if i != 0 {
+			children[i].SetPreviousID(childIDs[i-1])
+			children[i].SetFirstID(firstID)
+		}
+		if i == len(children)-1 {
+			children[i].SetParent(parent)
+		}
+		children[i].SetPayload([]byte{byte(i), byte(i + 1), byte(i + 2)})
+		children[i].SetPayloadSize(3)
+		childIDs[i] = children[i].GetID()
+		childAddrs[i] = object.AddressOf(children[i])
+	}
+
+	for i := range children {
+		require.NoError(t, e.Put(children[i], nil))
+	}
+
+	_, err = e.Get(parentAddr)
+	var splitErr *objectSDK.SplitInfoError
+	require.ErrorAs(t, err, &splitErr)
+	// Do not put link, so parent will have SplitInfo without link
+	require.True(t, splitErr.SplitInfo().GetLink().IsZero())
+	require.Equal(t, childIDs[len(childIDs)-1], splitErr.SplitInfo().GetLastPart())
+	require.Equal(t, childIDs[0], splitErr.SplitInfo().GetFirstPart())
+
+	tickEpoch(es, e)
+
+	require.Eventually(t, func() bool {
+		// Check this way because Get returns ErrObjectNotFound for expired error
+		for _, sh := range e.sortedShards(parentAddr) {
+			_, err = sh.Get(parentAddr, false)
+			if errors.Is(err, statusSDK.ErrObjectNotFound) {
+				return true
+			}
+		}
+		return false
+	}, 1*time.Second, 100*time.Millisecond)
 }
 
 func addExpirationAttribute(obj *objectSDK.Object, epoch uint64) {